just booking baki

+++ src/modules/Vehicles/vehicles.service.ts
@@ -28,21 +28,59 @@ const getSingleVehicles = async (vehicleId: string) => {
     return result.rows[0]
 }
 
-const updateVehicles = async(payload:Record<string,unknown>,id:string)=>{
+const updateVehicles = async (
+    payload: Record<string, unknown>,
+    id: string
+) => {
+    const {
+        vehicle_name = null,
+        type = null,
+        registration_number = null,
+        daily_rent_price = null,
+        availability_status = null
+    } = payload as any
 
-    const {vehicle_name, type,registration_number,daily_rent_price,availability_status	} = payload
-    const result = await pool.query(`
-        UPDATE vehicles SET vehicle_name = $1, type = $2, registration_number = $3, daily_rent_price =$4, availability_status = $5 WHERE id = $6 RETURNING *
-        `,[vehicle_name, type,registration_number,daily_rent_price,availability_status,id])
+    const result = await pool.query(
+        `
+        UPDATE vehicles
+        SET
+            vehicle_name = COALESCE($1, vehicle_name),
+            type = COALESCE($2, type),
+            registration_number = COALESCE($3, registration_number),
+            daily_rent_price = COALESCE($4, daily_rent_price),
+            availability_status = COALESCE($5, availability_status)
+        WHERE id = $6
+        RETURNING *
+        `,
+        [
+            vehicle_name,
+            type,
+            registration_number,
+            daily_rent_price,
+            availability_status,
+            id
+        ]
+    )
 
-        return result
-
-
-}
-const deleteVehicles = async(id:string)=>{
-    const result = pool.query(`DELETE FROM vehicles WHERE id = $1`,[id])
     return result
 }
+
+const deleteVehicles = async (id: string) => {
+  const bookings = await pool.query(
+    `SELECT * FROM bookings WHERE vehicle_id=$1`,
+    [id]
+  );
+
+  const booking = bookings.rows[0];
+
+  if (booking && (booking.status as string) === "active") {
+    return null;
+  }
+  const result = await pool.query(`DELETE FROM vehicles WHERE id=$1`, [id]);
+  return result;
+};
+
+
